unified-bench: run benchprof in-process
