+++ cmd/unified_bench/integration_test.go
@@ -2,7 +2,9 @@ package main
 
 import (
 	"bytes"
+	"os"
 	"os/exec"
+	"path/filepath"
 	"strings"
 	"testing"
 	"time"
@@ -66,4 +68,59 @@ func TestIntegration_NoHangLargeKeys(t *testing.T) {
 	}
 
 	t.Logf("Benchmark completed successfully.\nStdout:\n%s\nStderr:\n%s", stdout.String(), stderr.String())
+}
+
+func TestIntegration_ProfileDir_AutoRunsBenchprof(t *testing.T) {
+	if testing.Short() {
+		t.Skip("skipping integration test in short mode")
+	}
+
+	// Build the unified-bench executable first to test the CLI path.
+	buildCmd := exec.Command("go", "build", "-o", "../../bin/unified-bench", ".")
+	buildCmd.Dir = "."
+	if err := buildCmd.Run(); err != nil {
+		t.Fatalf("failed to build unified-bench: %v", err)
+	}
+
+	profileDir := t.TempDir()
+
+	cmd := exec.Command(
+		"../../bin/unified-bench",
+		"-test", "sequential_write",
+		"-dbs", "treedb",
+		"-keys", "128",
+		"-format", "table",
+		"-profile-dir", profileDir,
+		"-progress=false",
+	)
+
+	var stdout, stderr bytes.Buffer
+	cmd.Stdout = &stdout
+	cmd.Stderr = &stderr
+
+	timeout := 120 * time.Second
+	timer := time.AfterFunc(timeout, func() {
+		if cmd.Process != nil {
+			_ = cmd.Process.Kill()
+		}
+		t.Errorf("benchmark process timed out after %v", timeout)
+	})
+	defer timer.Stop()
+
+	if err := cmd.Run(); err != nil {
+		t.Fatalf("benchmark failed: %v\nStdout:\n%s\nStderr:\n%s", err, stdout.String(), stderr.String())
+	}
+
+	if _, err := os.Stat(filepath.Join(profileDir, "benchprof_results.json")); err != nil {
+		t.Fatalf("expected benchprof_results.json: %v", err)
+	}
+	if _, err := os.Stat(filepath.Join(profileDir, "insights.md")); err != nil {
+		t.Fatalf("expected generated insights.md: %v\nStdout:\n%s\nStderr:\n%s", err, stdout.String(), stderr.String())
+	}
+	if _, err := os.Stat(filepath.Join(profileDir, "insights.json")); err != nil {
+		t.Fatalf("expected generated insights.json: %v\nStdout:\n%s\nStderr:\n%s", err, stdout.String(), stderr.String())
+	}
+	if _, err := os.Stat(filepath.Join(profileDir, "insights.html")); err != nil {
+		t.Fatalf("expected generated insights.html: %v\nStdout:\n%s\nStderr:\n%s", err, stdout.String(), stderr.String())
+	}
 }
